Conditionally show login form.
Create a loader ui that is shown when the appvc control gets is-loading state.
We can listen for the ui request to show the login form to determine if we are done
loading and can show the login form.

In the future we might use a route instead of ui request event.